CRUD operations

+++ src/main/java/com/nous/app/repo/RegionRepository.java
@@ -8,10 +8,10 @@ import org.springframework.data.repository.CrudRepository;
 import com.nous.app.model.Region;
 
 
-	public interface RegionRepository extends CrudRepository<Region, Integer>{
+	public interface RegionRepository extends CrudRepository<Region, Long>{
 
 		public List<Region> findByRegionName(String regionName);
-		public List<Region> findByRegionId(int regionId);
+		public List<Region> findByRegionId(Long regionId);
